// src/components/common/Table/types.ts
import { Dispatch, SetStateAction } from 'react';
import { UseQueryResult } from 'react-query';
export type UseTableProps = {
  take: number;
  totalItemsByServer?: number;
  queryConfig: {
    queryFn: (props: any) => Promise<any>;
    key?: string;
  };
};

export type UseTableReturn<T = any> = {
  pagination: {
    pageIndex: number;
    onPageChange: (nextPage: number) => number;
    totalPages: number;
  };
  query: UseQueryResult<T, unknown>;
  setFilters: Dispatch<SetStateAction<Record<string, any>>>;
  setEnableRequest: Dispatch<SetStateAction<boolean>>;
  enableRequest: boolean;
};